+++ packages/entities/entities-vaults/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.17.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.17.1...@kong-ui-public/entities-vaults@3.17.2) (2025-04-24)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies (minor) ([#2088](https://github.com/Kong/public-ui-components/issues/2088)) ([81354c5](https://github.com/Kong/public-ui-components/commit/81354c5dc545cd9cc7fc830e5e34448c4f1ba7ae))
+
+
+
+
+
+## [3.17.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.17.0...@kong-ui-public/entities-vaults@3.17.1) (2025-04-23)
+
+
+### Bug Fixes
+
+* **typing:** update k-clipboard-provider related typing changes ([#2089](https://github.com/Kong/public-ui-components/issues/2089)) ([b7b46b4](https://github.com/Kong/public-ui-components/commit/b7b46b4fde7d2ae54235bfa9ea2b464ae35505d6))
+
+
+
+
+
 # [3.17.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-vaults@3.16.14...@kong-ui-public/entities-vaults@3.17.0) (2025-04-17)